fix: ScoreType String에서 Enum으로 변경
