+++ src/main/java/com/gloddy/server/core/event/score/ScoreUpdateEvent.java
@@ -1,6 +1,7 @@
 package com.gloddy.server.core.event.score;
 
 import com.gloddy.server.auth.entity.User;
+import com.gloddy.server.reliability.entity.vo.ScoreType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -10,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ScoreUpdateEvent {
     private User user;
-    private String ScoreType;
+    private ScoreType type;
 }
